# Faster Expression parser / Tokenizer with StringScanner