--- a/Gemfile
+++ b/Gemfile
@@ -24,3 +24,6 @@ group :test do
   gem 'rubocop-shopify', '~> 2.12.0', require: false
   gem 'rubocop-performance', require: false
 end
+
+gem "strscan", ">= 3.1"
+gem "lru_redux"