@ggmichaelgo 2025-01-23
> Following up, this is what we found:
> 
> https://shopify.slack.com/archives/C015H7QQNKD/p1737485547114139?thread_ts=1737132044.615309&cid=C015H7QQNKD
> 
> Michael Rokas I did some debugging with the help of [@timothy.willard](https://shopify.slack.com/team/W018HUJJ76X) and we have determined this is not a bug with the liquid gem, rather it is a bug with how we use the liquid gem. Our liquid helper that resolves usages [uses an ignorable_lookups](https://github.com/Shopify/flow-runtime/blob/bd9bc9ddd9c293aee600babf8429212d17c9c2a7/lib/maestro/util/liquid_helper.rb#L172) table to determine if a lookup should be ignored. This is used to not create usages resulting in hydration of assigns that are not actually used in downstream liquid. This table uses the entire ruby object as the key. This was working fine before because identical lookups were still separate ruby objects. But the 5.6.1 liquid gem introduces a cache which now reuses the same lookup object if the lookups are the same. This ends up resulting in our validator believing there is a direct usage of the second assigned variable which doesn't exist as a maestro variable only a liquid assign. There doesn't seem to be quick fix for this off hand, rather we will probably need to rework our logic for ignoring unused lookups to tell if the same lookup is used in different places since we can no longer rely on them being separate ruby objects.

Hello @plnii 👋 
For your use case, you could disable the Expression caching from the `ParseContext` like this:
```ruby
parse_context = Liquid::ParseContext.new(expression_cache: false)
template = Liquid::Template.parse(template, parse_context)
```